Move account module into separate file
